backend/document_uploader/requirements.txt
annotated-types==0.7.0
anyio==4.12.0
boto3==1.42.19
botocore==1.42.19
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jmespath==1.0.1
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==1.2.5
langchain-text-splitters==1.1.0
langsmith==0.5.2
nltk==3.9.2
orjson==3.11.5
packaging==25.0
pydantic==2.12.5
pydantic_core==2.41.5
PyMuPDF==1.26.7
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
s3transfer==0.16.0
six==1.17.0
tenacity==9.1.2
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
uuid_utils==0.12.0
zstandard==0.25.0